pyproject.toml
[project]
name = "wireguard"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "pillow>=11.3.0",
    "qrcode[pil]>=8.2",
]
